clean up bad files
