+++ stk/ops/eltwise_ops_test.py
@@ -1,6 +1,5 @@
 import unittest
 import itertools
-import numpy as np
 import torch
 from absl.testing import parameterized
 
@@ -15,7 +14,7 @@ def allclose(x, y, pct=0.25):
         return False
     return True
 
-_MUL_MATRIX_SIZES = (
+_MATRIX_SIZES = (
     (128, 128, 0.0),
     (256, 256, 0.5),
     (2048, 1024, 0.8),
@@ -37,15 +36,21 @@ _DTYPE = (
 )
 
 def _generate_testcases():
-    testcases = itertools.product(_MUL_MATRIX_SIZES, _DTYPE)
+    testcases = itertools.product(_MATRIX_SIZES, _DTYPE)
     testcases = [(*size, 128, dtype) for 
         (size, dtype) in testcases]
     return testcases
 
-_MUL_TEST = _generate_testcases()
+_ELTWISE_OP_TESTS = _generate_testcases()
 
+def _dense_and_sparse(rows, cols, sparsity, blocking, dtype, std=0.1):
+    mask = stk.random.dense_mask(rows, cols, sparsity, blocking)
+    dense = (torch.randn(rows, cols) * std * mask).type(dtype)
+    sparse = stk.ops.to_sparse(dense, blocking)
+    cuda_device = torch.device("cuda")
+    return (dense.to(cuda_device).requires_grad_(True),
+            sparse.to(cuda_device).requires_grad_(True))
 def _dense_and_sparse_like(x, std=0.1):
-
     dense_data = torch.randn_like(x.data) * std
     sparse = stk.Matrix(x.size(),
                         dense_data,
@@ -58,22 +63,12 @@ def _dense_and_sparse_like(x, std=0.1):
     return (dense.to(cuda_device).requires_grad_(True),
             sparse.to(cuda_device).requires_grad_(True))
 
-def _dense_and_sparse(rows, cols, sparsity, blocking, dtype, std=0.1):
-    mask = stk.random.dense_mask(rows, cols, sparsity, blocking)
-    dense = (torch.randn(rows, cols) * std * mask).type(dtype)
-    sparse = stk.ops.to_sparse(dense, blocking)
-    cuda_device = torch.device("cuda")
-    return (dense.to(cuda_device).requires_grad_(True),
-            sparse.to(cuda_device).requires_grad_(True))
 
+@parameterized.parameters(_ELTWISE_OP_TESTS)
+class EltwiseOpsTest(parameterized.TestCase):
 
+    def testElemwiseMul(self, m, n, sparsity, blocking, dtype):
 
-class LinearOpsTest(parameterized.TestCase):
-
-    @parameterized.parameters(_MUL_TEST)
-    def testElemwiseMulWithLike(self, m, n, sparsity, blocking, dtype):
-
-        # Common mask used to ensure same topology for a and b
         a_dense, a = _dense_and_sparse(m, n, sparsity, blocking, dtype)
         b_dense, b = _dense_and_sparse_like(a)
 
@@ -86,7 +81,6 @@ class LinearOpsTest(parameterized.TestCase):
 
         # Validate the results.
         out = stk.ops.to_dense(out)
-
         self.assertEqual(out.dim(), 2)
         self.assertEqual(expected_out.size()[0], out.size()[0])
         self.assertEqual(expected_out.size()[1], out.size()[1])
